requirements.txt
pinecone-client
transformers
pdfplumber
pandas
python-dotenv
tqdm
scikit-learn
